Change percent_discount to float

+++ db/migrate/20240229230701_create_bulk_discounts.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscounts < ActiveRecord::Migration[7.1]
   def change
     create_table :bulk_discounts do |t|
-      t.integer :percentage_discount
+      t.float :percentage_discount
       t.integer :quantity_threshold
       t.references :merchant, null: false, foreign_key: true
